pyproject.toml
[project]
name = "fastapi-boilerplate"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "croniter>=5.0.1",
    "fastapi>=0.115.0",
    "fastapi-pagination[sqlalchemy]>=0.15.3",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "h2>=4.3.0",
    "httpx-oauth>=0.16.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "tenacity>=9.0.0",
    "redis[hiredis]>=7.1.0",
    "sqlmodel>=0.0.27",
    "circuitbreaker>=2.0.0",
    "prometheus-client>=0.21.0",
    "uvicorn[standard]>=0.32.0",
    "alibabacloud_captcha20230305>=1.0.0",
    "alibabacloud_tea_openapi>=0.3.0",
    "alibabacloud_tea_util>=0.3.0",
    "openpyxl>=3.1.0",
    "playwright>=1.54.0",
    "chardet>=5.2.0",
    "funboost>=48.4.1",
]

[dependency-groups]
dev = [
    "fakeredis>=2.32.1",
    "pytest>=8.3.3",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.1.1",
    "pytest-timeout>=2.4.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.12.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
# filterwarnings = ["error::DeprecationWarning"]
timeout = 5
pythonpath = ["."]
testpaths = ["tests"]
addopts = ["--import-mode=importlib"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests",
]

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

